+++ views/login.handlebars
@@ -1,0 +1,3 @@
+<div class="container">
+  <a href="/auth/google">Login With Google</a>
+</div>
